% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal.loss.R
\name{causal.loss}
\alias{causal.loss}
\title{loss function that is used in outcome model estimation}
\usage{
causal.loss(y_true, y_pred)
}
\arguments{
\item{y_true}{two-column matrix of observed outcomes and exposure status}

\item{y_pred}{two-column matrix of predicted values for \eqn{\hat{\mu}_0} and \eqn{\hat{\tau}}}
}
\value{
loss value for predicted \eqn{\hat{\mu}_0} and \eqn{\hat{\tau}}, considering the loss function introduced in the reference.
}
\description{
loss function that is used in outcome model estimation
}
\references{
Max H. Farrell and Tengyuan Liang and Sanjog Misra (2021) \emph{Deep Neural Networks for Estimation and Inference}. The Econometric Society.
}
