% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal.loss.R
\name{loss_sum_square_error}
\alias{loss_sum_square_error}
\title{loss function sum of squars of errors}
\usage{
loss_sum_square_error(y_true, y_pred)
}
\arguments{
\item{y_true}{true labels}

\item{y_pred}{predicted vector}
}
\value{
sum of squars of errors.
}
\description{
loss function sum of squars of errors
}
