% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal.loss.R
\name{metric_mean_absolute_percentage_error2}
\alias{metric_mean_absolute_percentage_error2}
\title{the metric of mean of absolute percentage error}
\usage{
metric_mean_absolute_percentage_error2(y_true, y_pred)
}
\arguments{
\item{y_true}{two-column matrix of observed outcomes and exposure status}

\item{y_pred}{two-column matrix of predicted values for \eqn{\hat{\mu}_0} and \eqn{\hat{\tau}}}
}
\value{

}
\description{
the metric of mean of absolute percentage error
}
