% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal.loss.R
\name{causal.loss}
\alias{causal.loss}
\title{loss function that is used in outcome model estimations}
\usage{
causal.loss(y_true, y_pred)
}
\arguments{
\item{y_true}{real labels}

\item{y_pred}{predicted labels}
}
\value{
loss function of \eqn{\hat{\mu}_0} and \eqn{\hat{\tau}}
}
\description{
loss function that is used in outcome model estimations
}
